fix: modal.tsx supabase CSR 로직 분리 및 util 로직 분리
